edits after discussing with jay
