+++ test/python/ibmq/test_ibmq_integration.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2017, IBM.
+# Copyright 2018, IBM.
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
@@ -8,19 +8,17 @@
 # pylint: disable=redefined-builtin
 
 
-"""Compiler Test."""
+"""IBMQ provider integration tests (compile and run)."""
 
 import unittest
 
-import qiskit
+from qiskit import IBMQ
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
 from qiskit.transpiler import transpile
-from qiskit import compile
+from qiskit import compile, execute
 from qiskit.result import Result
-from qiskit import execute
 from qiskit.providers.ibmq import least_busy
-from ..common import QiskitTestCase
-from ..common import requires_qe_access
+from ..common import QiskitTestCase, requires_qe_access
 
 
 class TestCompiler(QiskitTestCase):
@@ -28,14 +26,31 @@ class TestCompiler(QiskitTestCase):
 
     seed = 42
 
+    def setUp(self):
+        qr = QuantumRegister(1)
+        cr = ClassicalRegister(1)
+        self._qc1 = QuantumCircuit(qr, cr, name='qc1')
+        self._qc2 = QuantumCircuit(qr, cr, name='qc2')
+        self._qc1.measure(qr[0], cr[0])
+        self._qc2.x(qr[0])
+        self._qc2.measure(qr[0], cr[0])
+
+    @requires_qe_access
+    def test_ibmq_result_fields(self, qe_token, qe_url):
+        """Test components of a result from a remote simulator."""
+        IBMQ.enable_account(qe_token, qe_url)
+        remote_backend = IBMQ.get_backend(local=False, simulator=True)
+        remote_result = execute(self._qc1, remote_backend).result()
+        self.assertEqual(remote_result.backend_name, remote_backend.name())
+        self.assertIsInstance(remote_result.job_id, str)
+        self.assertEqual(remote_result.status, 'COMPLETED')
+        self.assertEqual(remote_result.results[0].status, 'DONE')
+
     @requires_qe_access
     def test_compile_remote(self, qe_token, qe_url):
-        """Test Compiler remote.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = least_busy(qiskit.IBMQ.backends())
+        """Test Compiler remote."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = least_busy(IBMQ.backends())
 
         qubit_reg = QuantumRegister(2, name='q')
         clbit_reg = ClassicalRegister(2, name='c')
@@ -49,12 +64,9 @@ class TestCompiler(QiskitTestCase):
 
     @requires_qe_access
     def test_compile_two_remote(self, qe_token, qe_url):
-        """Test Compiler remote on two circuits.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = least_busy(qiskit.IBMQ.backends())
+        """Test Compiler remote on two circuits."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = least_busy(IBMQ.backends())
 
         qubit_reg = QuantumRegister(2, name='q')
         clbit_reg = ClassicalRegister(2, name='c')
@@ -70,12 +82,9 @@ class TestCompiler(QiskitTestCase):
 
     @requires_qe_access
     def test_compile_run_remote(self, qe_token, qe_url):
-        """Test Compiler and run remote.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = qiskit.IBMQ.get_backend(local=False, simulator=True)
+        """Test Compiler and run remote."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = IBMQ.get_backend(local=False, simulator=True)
 
         qubit_reg = QuantumRegister(2, name='q')
         clbit_reg = ClassicalRegister(2, name='c')
@@ -90,12 +99,9 @@ class TestCompiler(QiskitTestCase):
 
     @requires_qe_access
     def test_compile_two_run_remote(self, qe_token, qe_url):
-        """Test Compiler and run two circuits.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = qiskit.IBMQ.get_backend(local=False, simulator=True)
+        """Test Compiler and run two circuits."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = IBMQ.get_backend(local=False, simulator=True)
 
         qubit_reg = QuantumRegister(2, name='q')
         clbit_reg = ClassicalRegister(2, name='c')
@@ -112,12 +118,9 @@ class TestCompiler(QiskitTestCase):
 
     @requires_qe_access
     def test_execute_remote(self, qe_token, qe_url):
-        """Test Execute remote.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = qiskit.IBMQ.get_backend(local=False, simulator=True)
+        """Test Execute remote."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = IBMQ.get_backend(local=False, simulator=True)
 
         qubit_reg = QuantumRegister(2)
         clbit_reg = ClassicalRegister(2)
@@ -132,12 +135,9 @@ class TestCompiler(QiskitTestCase):
 
     @requires_qe_access
     def test_execute_two_remote(self, qe_token, qe_url):
-        """Test execute two remote.
-
-        If all correct some should exists.
-        """
-        qiskit.IBMQ.enable_account(qe_token, qe_url)
-        backend = qiskit.IBMQ.get_backend(local=False, simulator=True)
+        """Test execute two remote."""
+        IBMQ.enable_account(qe_token, qe_url)
+        backend = IBMQ.get_backend(local=False, simulator=True)
 
         qubit_reg = QuantumRegister(2)
         clbit_reg = ClassicalRegister(2)
