+++ test/python/result/__init__.py
@@ -5,4 +5,4 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Qiskit Result"""
+"""Qiskit Result tests."""
